Visual changes in the timetable in case of non-default schedule
This PR indroduces new feature to zorza. From now alternative periods will be shown if that day's lessons have different than default schedule. Additionally, cancelled lessons will be crossed out of the timetable.

* showing alternative periods and edit of css when lessons are cancelled

* do not strike through cancelled lessons

* Treat the possibility of multiple alternative schedules.

* Show alternative periods only when there is a non-default day plan.

* Show only one alternative period
Show alternative period that corresponds to current DayPlan

* Handle cancelled lessons in dayplan.
When the lesson is cancelled in current dayplan,
strikethrough the original time.
Optimize query on DayPlan objects.

* Handle cancelled lessons in dayplan.
When the lesson is cancelled in current dayplan,
strikethrough the original time.
Optimize queries on DayPlan and alternative_periods.

* Treat the case when lessons are cancelled
for a whole day

* Fix bugs and clean code

---------

Co-authored-by: Zuzana Lüköová <zuzana.lukoova@student.tuke.sk>
Co-authored-by: Ignacy Świderski <ignacyswiderski@wp.pl>